Core: split glyph rasterization out of the font registry
