+++ takumi-core/src/resources/mod.rs
@@ -1,5 +1,7 @@
 /// Font loading and processing functionality
 pub mod font;
+/// Glyph rasterization: shaped glyph ids to bitmaps or vector outlines.
+pub mod glyph;
 /// Image state and resource management
 pub mod image;
 /// Backend-agnostic decoded-image buffer
